backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.27
psycopg[binary]>=3.2
pgvector>=0.4
sentence-transformers>=3.0
numpy>=1.26
python-dotenv>=1.0
pydantic>=2.6
